+++ web/modules/custom/retraitespopulaires/rp_site/src/Form/SearchForm.php
@@ -53,6 +53,9 @@ class SearchForm extends FormBase {
         $form['#action'] = Url::fromRoute('rp_site.search')->toString();
         $form['#method'] = 'GET';
 
+        $form['#token'] = FALSE;
+        $form['#build_id'] = FALSE;
+
         $form['input-group_start'] = array(
             '#markup' => '<div class="input-group full-width">',
         );
@@ -67,10 +70,11 @@ class SearchForm extends FormBase {
         );
 
         $form['actions']['submit'] = array(
-            '#type'        => 'submit',
-            '#attributes'  => ['class' => array('btn btn-default'), 'icon' => 'retraitespopulaires-icon retraitespopulaires-icon-search'],
-            '#prefix'      => '<span class="input-group-btn">',
-            '#suffix'      => '</span>',
+           '#id'          => 'rp_site_search_form_submit',
+            '#type'       => 'submit',
+            '#attributes' => ['class' => array('btn btn-default'), 'icon' => 'retraitespopulaires-icon retraitespopulaires-icon-search'],
+            '#prefix'     => '<span class="input-group-btn">',
+            '#suffix'     => '</span>',
         );
 
         $form['input-group_end'] = array(